+++ app-schematics/src/rxjs-page/index.ts
@@ -3,7 +3,7 @@ import { classify, dasherize } from '@angular-devkit/core/src/utils/strings';
 import { Rule, SchematicContext, Tree, apply, mergeWith, move, template, url } from '@angular-devkit/schematics';
 import { Schema } from './schema';
 
-const RXJS_ROUTING_MODULE_PATH = './src/app/rxjs-layout-page/rxjs-layout-page-routing.module.ts';
+const RXJS_ROUTING_MODULE_PATH = './src/app/comprende-rxjs-layout/comprende-rxjs-layout-routing.module.ts';
 
 export function rxjsPage({ name }: Schema): Rule {
   return (tree: Tree, _context: SchematicContext) => {
@@ -15,7 +15,7 @@ export function rxjsPage({ name }: Schema): Rule {
         date: `${new Date().getDate()} de ${new Date().toLocaleDateString('es-ES', { month: 'long' })} de ${new Date().getFullYear()}`,
         ...strings,
       }),
-      move('src/app'),
+      move('src/app/comprende-rxjs-layout'),
     ]);
 
     // add new path to the component in the routing module
@@ -24,7 +24,7 @@ export function rxjsPage({ name }: Schema): Rule {
     rxjsRoutingModule = `${rxjsRoutingModule.slice(0, indexToInsert)}
       {
         path: '${name}',
-        loadChildren: () => import('../${dasherize(name)}/${dasherize(name)}.module').then((m) => m.${classify(name)}Module),
+        loadChildren: () => import('./${dasherize(name)}/${dasherize(name)}.module').then((m) => m.${classify(name)}Module),
       },${rxjsRoutingModule.slice(indexToInsert)}`;
     tree.overwrite(RXJS_ROUTING_MODULE_PATH, rxjsRoutingModule);