+++ src/app/app.component.ts
@@ -1,6 +1,7 @@
-import { Component } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { Component } from '@angular/core';
 import { RouterOutlet } from '@angular/router';
+import { SeoService } from '@jaimemartinmartin15/jei-devkit-angular-shared';
 
 @Component({
   selector: 'app-root',
@@ -9,4 +10,8 @@ import { RouterOutlet } from '@angular/router';
   template: `<router-outlet></router-outlet>`,
   styles: [':host { display: block }'],
 })
-export class AppComponent {}
+export class AppComponent {
+  public constructor(readonly seoService: SeoService) {
+    seoService.listenNavigationEvents();
+  }
+}