+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## 1.0.1
+
+Use jei-devkit-angular-shared.
+Fix seo service after migration.
+
 ## 1.0.0
 
 Copied code from [main repository](https://github.com/jaimemartinmartin15/jaimeelingeniero.es) and adapt it to the new repo.